extract_unity53_shaders: Disassemble DX9 shaders
This uses cmd_Decompiler as a convenient way to disassemble the DX9
version of the shaders. fxc is a good alternative, but would require the
Windows SDK to be downloaded for most people to obtain it. Modifying
shaderasm.exe to disassemble as well would also work, but seems
pointless since cmd_Decompiler can already do it and is required for the
DX11 shaders anyway.