Unity5 / TLD / hlsltool: Add support for the new directional shadow shader
Unity 5.3 adds a new output to the directional lighting shader, changing
the hash and messing up the scripts. Add support to override the FOV
texcoord in hlsltool and use it to fix directional lighting in The Long
Dark. Add the manual fixes to the vertex shader and d3dx.ini.