The Long Dark: Fix footprint decals
Similar pattern to lighting. For now using a view-space correction, but
the FOV is not yet obtained in the frame, so disabled clearing it on
present and using stale data from the previous frame. Not really
noticeable, but if it's a problem it's probably possible to do a
world-space correction instead.