hlsltool: Add fallback path to lighting shaders to use a WS correction
This adds a fallback path to the lighting pixel shaders to use a world
space correction whenever the FOV is not available. This should
generally make the script a little easier to use as it reduces the
chance of a complete failure.

In Stranded Deep I see the shadow adjustment break periodically - need
to investigate.