Game List: Reduce API query to 250/page
This seems like it might have some success resolving the outage on the
game list. Bo3b collected stats showing that the query was taking 5
seconds in some regions and timing out, limiting it to 250/query and
using the pagination API to obtain the rest will hopefully eliminate the
timeouts.