Shader database: Use itertools.count() instead of range(3)
Pretty sure the range(3) was something I used to make sure the script
could not run away, but since it seemed reliable remove the limit.

Also reduce the results/page to 50 in case it hits a similar issue to
the game list (especially as it requests more data than the game list,
which could increase the time).