Game list: Remove healthy heartbeat ping, leave error ping
The issues seem to be resolved, so it is not all that interesting to see
the ok pings anymore. Leave the error ping in place, since it usually
won't do anything, but if there is another problem in the future it
provides much more useful debugging info than the API console does.