extract_unity53_shaders: Add --skip-classic-headers option
This synthesizes the sub_program and related objects expected by the
classic extract_unity_shaders script so that it can be used on a
.shader.decompressed file if the corresponding .shader file is not
available.

Not recommended for general use since the .shader file contains
important information regarding shadow casters that will not be
extracted without it. This is intended to be used in the event that
someone provides a .shader.decompressed file that the script fails on
without the corresponding .shader file.