The Park: Fix UE4 ReflectionEnvironmentComputeShaders.ucf shader
The solution was fairly simple as it turns out, but I must have made a
typo earlier and ended up chasing a red herring. I believe this is a
standard shader in UE4, and was able to match up some of the code &
structures.

Note that the source code indicates that this is a TILED REFLECTION
shader, so it may still be broken in other instances.