The Park: Fix UE4 tile lighting
Simplest tile lighting yet - everything we need is in the one shader
(for once)! The trick is simply to apply a correction to the light
bounds where it is computing which tiles get which lights.

Even easier since we have the source code - fairly easy to identify the
correct usf shader & corresponding cpp source :)