The Park: Use resource copying instead of fake_o0
fake_o0 does not cache it's dummy render target, so is not suitable for
shaders used multiple times per frame and can quickly kill the frame
rate, and appears that it might have a memory leak. Use resource copying
instead, which will cache the render targets and has a negligable impact
on frame rate.