Witcher 3: Replace fake_o0 with resource copying
fake_o0 is reported to cause issues with some shaders, due to the fact
that it does not cache or reuse render targets and can cause havoc in
vram management if it is used multiple times per frame.

This is not one of those cases, but modern 3DMigoto has a superior
alternative to fake_o0 with arbitrary resource copying, so switch over
to using that.