Add inverse-cs.hlsl to inverse a matrix from a custom shader
This custom shader may be used to inverse a matrix rather than doing it
inline. This is particularly useful for situations where an inverse
needs to be calculated in an assembly shader, which is complicated
(though not impossible) to do inline.

This requires 3DMigoto 1.2.40, which fixes several issues creating
custom non-structured buffers.

This approach calculates the inverse in parallel using four threads on
the GPU, which may be slightly faster than calculating the full inverse
in every thread as we do when calculating it inline. Of course, this
approach also costs a full dispatch call, which incurs a cost in itself
(possibly more so if the GPU has to switch from graphics to compute
pipelines) so without measuring the performance it is impossible to say
which approach is faster.

https://forums.geforce.com/default/topic/685657/3d-vision/3dmigoto-now-open-source-/post/4909267/#4909267