3dvisionlive.js: Fix distance method if a screenshot is duplicated
The distance viewing method does some post processing and looks up the
image by an ID, but that ID may be reused if the same screenshot is
included on the page multiple times.

Add the block number to the ID to keep it unique between separate blocks
(and there isn't much point repeating an image in the same block, so I
don't care about that).