3dvisionlive.js: Link images to full size download
Clicking on a cross-eyed, distance or anaglyph screenshot will open the
full size version in a new window. Distance still opens the cross-eyed
version since at that size it won't be possible to use the distance
viewing method and is more important to just save it as a .jps to open
in a dedicated viewer.