cleanup_unity_shaders: Fix adding ini file when game is symlinked
'Stranded Deep' is currently a symlink to 'Stranded Deep (DX11)', which
causes git add to bail out if the full path to the d3dx.ini is used.
Since we already pass the -C flag to git to set the working directory we
don't need to use the full path, so just use the relative path instead
which works just fine.