CustomSettingNames: Add alternate IDs for several Stereo settings
It seems the 3D Vision driver uses several alternate IDs for some of the
settings, possibly for backwards compatibility? The mapping of these IDs
can be found in two tables above the table with the names, but a given
game will only use one of the tables, so only one of the two possible
alternate IDs may work in a given game. These alternate IDs do seem to
take priority over the primary ID.