sanitise_nv_profiles: Properly decode internal settings
This now properly decodes all the settings instead of just sanitising
the encoding. The produced files are now back in utf16 since some of the
strings do legitimately contain non-ascii characters.

The internal settings (both strings and DWORDs) are XORed with a 2048
bit key offset with a value derived from the setting ID. The key can be
recovered using plain text attack or other methods, and the offsets are
relatively straight forward to determine.