asmtool: Add --fix-unity-reflection and --fix-unity-frustum-world
This adds assembly versions of these two fixes, which will allow the
autofix to use asmtool instead of hlsltool for the large number of
generic fixes where decompiler bugs are a risk, reserving hlsltool for
use on the more specific fixes that the decompiler is known to work well
with.

Since these fixes require the MVP matrix to be inverted, which is
difficult to do in hand crafted assembly, they include a dedicated
shader to invert this matrix and modifications to the template to allow
it's use.