hlsltool/asmtool: Take into account alternate filenames when overwriting a shader
We may mix HLSL and asm shaders in a fix, but we don't want to have both
asm and HLSL versions of a single shader present, as that can become
confusing as to which will be used.

Have both these tools check for both HLSL and asm versions of the
filename (as well as ~failed for HLSL and ~bad for both), refusing to
install a shader if any of these are present. --force will actively
remove alternate version of the shader (except for ~bad variants) since
we cannot rely on implicitly overwriting them if their filenames differ.