cleaup_unity_shaders: Add --remove-ini-sections-for-uninstalled-shaders option
This option will remove ini sections for any shader that is not
installed in ShaderFixes/ShaderOverride. It is intended to be used if
the shaders are to be regenerated and their sections need to be
regenerated as well.

It might also be useful to cleanup ini sections for shaders in non-Unity
games.

Take caution - unlike Helix Mod, 3DMigoto does not require a shader to
be installed in order to add a section to the ini for it, and this
option would remove all such sections.