Unity5 DX11: Switch halo, reflection and frustum fixes to assembly
These fixes are run against just about every shader found in a game,
which drastically increases the chance of a decompiler bug affecting
some of them causing failed compilations or rendering issues.

Switch them to use asmtool with the assembly versions of the shader to
eliminate these issues. As a bonus, this should reduce the first launch
loading time since asm shaders take significantly less time to assemble
than HLSL shaders take to compile.

The script may still break the odd shader, but this should reduce the
number that are incidentally broken by external forces.

Arguably the frustum fix probably didn't need to be switched since it
only affects a small number of shaders in most games, but since it is
done in the same pass as the vertex halo and camera position fixes it
makes sense to switch it as well.