sanitise_nv_profiles: Handle corrupt string in Battlefield: Hardline profile
One of the strings in the Battlefield Hardline profile appears to be
corrupt as it contains a NULL character at the start of the string
preceding other data. Decrypting this string prevents the file from
being re-imported into Geforce Profile Manager, but leaving it encrypted
violates the utf16 encoding of the overall file.

Detect this specific case and replace it with a hex representation of
the encrypted and decrypted strings.