MGSVTPP: Fix throwable item line
The shader for the shimmer also is responsible for the line showing
where a grenade will be tossed. We had undone the driver's correction
for the shimmer, but we want the line corrected. Adjust when w != ~1 to
work for both.

This might be a side effect of changing the hashes for MGO...