matrix.hlsl: Add macro to make matrix from 16 consecutive constants
It is common that we will not have a definition of a matrix, but will
have identified one in a constant buffer. This macro adds a more
convenient means of creating a matrix variable out of one by only
specifying which constant buffer and the starting offset of the matrix.