extract_unity53_shaders: Extract separate shaders with same hash and merge headers
This makes the script behave like the old extract_unity_shaders script,
delaying writing out the shaders until it has finished parsing all that
it was asked to do, and merging their headers together.

The shader will be extracted to all relevant shader directories, instead
of stopping after the first. This is important as the lighting vertex
shaders tend to all have the same hash, and the autofix script depends
on locating pixel shaders by that hash - if the shader is only extracted
to the first lighting shader directory, the autofix will miss the
subsequent ones.

This has become more important in Unity 5.4 where PrePassLighting and
DeferredReflections share the same vertex shader, but
DeferredReflections is extracted first preventing the fix from picking
up PrePassLighting and missing all the point & spot lights. In previous
versions of Unity 5.x this was less of an issue as the point and spot
lights were all handled by DeferredShading, which was extracted first
(though could have been an issue for games using the legacy Unity 4.x
lighting path).