extract_unity53_shaders: Output 5.3 header section and merge *all* headers
Reusing the code from the old script doesn't quite do the right thing as
it misses the new 5.3 header section. It also complicates the output
since it creates a number of sections for each distinct shader, but
these would need to be separate again from the 5.3 sections...

Rather than creating separate sections for each distinct shader, just
merge all the headers together, which will always result in two sections
- one for legacy headers and one for the 5.3 headers with the usual
markings to indicate which lines came from where.

Originally I did not do this as there were cases where this created a
complicated diff for no good reason if the unity headers differed too
much between identical shaders. I'll have to keep an eye out and see if
this is still the case, but given the legacy headers are much simpler we
might get lucky and avoid it.