extract_unity53_headers: Improve merged headers by deprioritising scope changes
Since we are now merging headers from different shaders (so long as
their bytecode is identical), the diffs can get a little complicated if
the unity headers differ too much. Deprioritise outputting headers that
include a scope change in the merge strategy to increase the chance of
outputting like headers together simplifying the diff.