Firewatch: Accurate fog fix
The fog on the main menu was accurately fixed, but in game it was only
fixed in certain directions and broken in others. This seems to be an
issue that the copied matrices are for a different frame of reference -
I may need to investigate that aspect in more detail later.

For now, the easiest solution is to use the camera frustum from the
vertex shader to calculate the direction of screen right in world space
and use it to always apply the adjustment in the correct direction.

This moves the fog adjustment entirely into the pixel shader to avoid
needing to adjust the world space position twice.