GG: Add UI adjustment from mouse cursor position
This adjusts all UI elements to depth around the mouse cursor.

Note that since it copies the depth texture from the shadow shaders,
they will no longer be able to be fixed by PSscript.lua, as they must
exist in the ShaderOverride directory for the relevant ini section to
take effect.

There are some issues in some of the menus where some items don't line
up or are clipped, possibly caused by one of the shaders (FE0DB3C7)
being used to render UI elements to an off-screen buffer.