GG: Fix shadow pixel shaders with shadertool
The autofix didn't seem to work before, but it turns out that was an
issue in another shader that PSscript.lua has also fixed. With that left
in place, shadertool can fix the shadow shaders just fine.

This is necessary to use these to copy the depth buffer for the UI
adjustment, as leaving them to PSscript.lua to fix will not allow the
ini sections to run.