GG: Disable DumpAll and remove PSscript.lua to fix crash
The game crashes pretty consistently when loading the Stairs level in
the Maya run - it seems to be related to disassembling out some
particular shader, which will always happen if either PSscript.lua or
DumpAll is in use.

Remove the script and kill the DumpAll option to resolve the crash, but
this means that we will need to fix some remaining issues by hand.