GG: Add first vertex hud adjustment option and make selectable
This adds a mode where the HUD will adjust from the position of the
first vertex in a given element. The result is similar to that of
adjusting each individual vertex, but each icon will be parallel to the
screen (however note that some UI elements (e.g. text) are made up of
multiple icons each of which could be at a different depth).

There is an attempt commented out to only use the first vertex position
of the crosshair for the whole UI, but it didn't seem to work properly -
I think the crosshair texture is shared with other things.

Also adds the old center screen adjustment for completeness, and adds
key bindings to switch between all available options.