shadertool: Add --fix-unreal-ps-halo pattern
This pattern is for certain Unreal Engine 3 games that include a VPM
multiplication in their pixel shaders and do not already include the
usual UE3 NVStereo code.

This also adds helper routines to locate row major matrix multiplies and
follow a component through a swizzle and a mask to locate the
corresponding component in the output register. The hlsl_swizzle routine
has also been moved from asmtool to shadertool.