GG: Run auto vertex halo fix
This fixes some effects such as the hearts on the background of doki
doki mode. These effects were fixed by PSscript.lua by adjusting the
input that was used to sample ScreenColorTexture, but at first glance
this seems to work just as well.

Note that the --fix-unreal-ps-halo is still necessary as it fixes issues
created entirely in the pixel shader, whereas the halos adjusted here
can be fixed in either.