shadertool: add --fix-unreal-halo-sct pattern (and rename other halo option)
This pattern fixes uses of SceneColorTexture in certain UE3 games. Like
the other UE3 halo fix, it should only be used if the game does not
already include the NvStereo halo code (might be something I can
detect).

These halos are rare for UE3 in that they can be fixed in the vertex
shader using the generic pattern, but doing so would break other
effects, so fixing them in the pixel shader is a better idea.