extract_unreal_shaders: UE4 WIP
This is a work in progress to extract shaders (or more to the point -
any available metadata) from UE4 games. With any luck this might
eventually lead to being able to script some patterns, but it isn't
there yet.

I'm somewhat annoyed that looking at the UE4 source code shows that it
does not have a well defined format and the format is mostly an ad-hoc
serialisation of their class objects, and there are no flags to indicate
which optional sections are present. This could easily lead to 3rd party
shaders having arbitrary data of an arbitrary length with no well
defined way to determine how long that is (to skip over it and get back
to the generic info we are interested in) without access to their source
code.