extract_unreal_shaders: Successfully extract shaders from ABZU
UE4 adds a tail to the shaders that we need to strip off before we can
write out the shaders. The format doesn't seem to match the code I'm
following, but that might just be due to different engine versions.

This now extracts the shaders from ABZU's GlobalShaderCache using the
embedded hash (i.e. *not* the normal 3DMigoto hash... not yet anyway).
It currently cannot determine what type of shader is being extracted, so
leaves a XX in the filename instead.