extract_unreal_shaders: uasset material shader map decoding WIP (the "are you fucking serious" rant edition)
The more I dig into this file format... the more I hate it. It is better
than their global shader cache since they at least made an attempt to
version it, but the thing had changed ad-hoc as developers have changed
the serialization code willy nilly, often forgetting to bump the version.

I can see the elegance in their serialization code, or at least what a
C++ developer would most certainly perceive as being elegant, but
seriously - fuck that shit and *design* me a good file format please.

If you are going to do something ad-hoc like this, you need to add
static testing that fails the build if it is modified without bumping
the version. e.g. a build script that validates the size of data that
each Serialize() and operator<<() function produce for *every* class
that could ever conceivably be put on disk, or called from some other
class' Serialize() method. Of course that would be easier if they were,
oh I dunno... structs!

And for god's sake, if you are going to have variable length data
structures, you need to provide some way for a parser to skip over them
without having to decode them. To be fair, there is some of this in the
top level uasset format, suggesting that they have thought about these
types of issues in the past, but digging down into the contained info is
where the format falls over - we need more detail than the top level
format allows us to find easily, but to get it we need to decode
*everything*.

It also has a horrible design where a class (and related serialization
code) is instantiated based on matching a name... which leaves the
format open to being game specific.

Despite not having the Unity source code... right now I'd have to say I
much prefer their approach at designing a file format, though it is
certainly not without issues either (and they also neglected to bump
their file version fields sometimes, but at least they have an engine
version field... Oh yeah, so does UE4 uasset files - except it isn't
fucking filled out! Engine version 0.0.0 my arse!).

I could keep digging... in theory I'm not that far from being able to
pull actual shaders and metadata out of this thing, but I've already
spent enough time on this that I could have used to fix several UE4
games, and I'm concerned that the maintenance cost of this script will
turn out to be enormous - and I'm just not sure if it is worth
continuing.

I've been considering for a while whether it would be worthwhile
switching to using the UE4 code itself to parse this, since it already
handles a lot of the edge cases and... *some* of the version
differences, but it has become painfully clear that it would fail to
even parse fairly recent versions of the uasset format since they have
neglected to bump the file version a number of times, and lack a number
of version checks that would be required to parse older files - the
parser would have to be rebuilt for each specific engine version
(similar problem with writing a plugin for this engine). Certainly
doable, and perhaps less maintenance cost than maintaining a separate
script, but this thing has me jaded. It is also not clear if that would
even work, since parts of the format seem to depend on third party
code that is not distributed with the base UE4 engine.

The asset parsing is disabled, since it currently skips a hardcoded
amount of data to find the start of the shader map, and that amount was
selected based on one specific shader I am using as an example. The
length is variable, and I have yet to identify the code responsible for
it to determine the length... and have some fear that it might turn out
to be shader/material and perhaps even game specific.

It may be worthwhile considering switching to a "dumb" script that just
scans for the "DXBC" signature and tries to pull a little of the
surrounding metadata out, rather than trying to parse the whole format -
that might be simpler and less specific to a version of the engine/game,
but wouldn't produce as much useful info... of course, this endeavour has
yet to prove if the info I'm getting will be useful for us at all.

If I continue working on this script - it will be out of stubbornness,
not love. But I can be pretty stubborn...