extract_unity_shaders: Add support for 1x4 matrices (fixes Yesterday Origins)
The AVPro* shaders reference the first row of UNITY_MATRIX_V, which for
DX9 causes Unity to only pass a single vector, which is a combination
that is not handled by my extraction scripts, triggering one of the
assertions.

This seems to confirm that "type3" signifies a matrix, and the
size1/size2 fields represent the matrix dimensions. Restructure the code
to check type3 before the sizes and relax the partial matrix cases to
accept a size of 1.

* Trial available here: http://renderheads.com/product/avpro-video/

Reported-by: Blacksmith56