Unity5 DX9-New autofix53: Use ????????.txt as wildcard instead of *.txt
If the extract_unity53_shaders script is run at any point without
cmd_Decompiler, it will generate *_headers.txt files instead of
attaching them to the assembly file. If these files are still present
later when the autofix script it run it will attempt to apply fixes to
them, which will fail and cause it to error out prematurely.

Of course cmd_Decompiler should be present from the get go so this isn't
a typical situation, but we can use a more explicit wildcard to avoid it
just in case.