Witcher 3: Make Helifax' 1.22 version master and remove other versions
It is confusing having three versions of the Witcher 3 fix existing side
by side in the git tree, when they are already tracked by the git
history and released zip files. Remove the two older versions and make
the latest update from Helifax the current master copy.

This also removes the reasm files and precompiled DLLs.