Witcher3: Remove HBAO+ toggle key
This toggle was originally added as a workaround for a driver bug to
keep the game playable for a while until the user could restart the
game. AFAIK that bug is long since fixed and we want the F11 key for the
SBS shader, so remove this. Still ship the vanilla shader for people who
don't know how to use uninstall.bat.