FCPrimal: Downscale and limit reflections to improve SLI framerate
Downscale the reflections to half resolution prior to performing
stereo2mono to drastically improve performance in SLI - this works best
when combined with reducing the water quality to 'High' in the settings.

Also limit the reflection downscaling & stereo2mono to the first time it
is encountered in a frame. Frame analysis showed that this was easily
occurring 14 times in a frame otherwise, needlessly wasting time
repeating the same operations.