Stranded Deep: Tweaks to hopefully reduce flickering
The fixed effects (noticeable on shadows, water, sand, reflections)
flicker on and off on some frames, but it is hard to catch the bad
frames in frame analysis to work out what is different about them. Make
several tweaks to the resource copying to try to reduce this as much as
possible:

- Require 3DMigoto 1.2.51 since it will have a bug fix for the max
copies/executions per frame

- Change all max copies/executions per frame to 1 to only get the first
use per frame

- Do not NULL out the copied resources on the present call. This will
allow stale information to be used, but given the FOV doesn't really
change much in this game it should generally be fine.