Stranded Deep: Manual halo fix for glass shader with unusual halo pattern
This shader just has a halo, but the pattern is slightly unusual since
the output position is not used directly in a texcoord, but rather the
version of it just before adding the final row of the MVP matrix is
used:

    add o0.xyzw, r0.xyzw, cb3[3].xyzw
    mad r0.xyzw, cb3[3].xyzw, v0.wwww, r0.xyzw

That's unusual enough that my scripts can't autofix it, so manually fix
it. The final row has to be added before applying the halo fix and
subtracted again after.