Stranded Deep: Fix Ceto water misalignment
This is the assembly equivalent of bd8221b71ca84feb3f7542f81986aeffe549273a.

Really need to script this one, but not positive how to reliably find
which texcoord has the depth value to do so, and make sure the
adjustment is inserted before the 2nd IVP multiply.

As before, only changing the shader that the game is using since I can't
validate the rest.