New script: generic_shader_extractor
This script aims to extract shaders from almost any DX11 game. Rather
than attempting to understand the game files like the Unity extraction
scripts it just searches all game files for the DXBC signature and a
valid MD5 like hash indicating the presence of a shader.

It does not yet determine what kind of shader is being extracted, so the
filename contains xx in place of the shader type.

Support for all three types of hashes are available, and is selected
with '--hash embedded', etc.