generic_shader_extractor/dx11shaderanalyse: Parse shader model
This parses the first part of the SHDR/SHEX bytecode sections to find
the shader type and model. dx11shaderanalyse will print it along with
the rest of the info it extracts, and generic_shader_extractor will use
it to create the correct filename for the shader type.

This does not yet handle exotic shader types such as DX9 feature levels
or libraries. There is some protection so these will probably continue
using 'xx' in the filename, but this is untested.