extract_unreal_shaders: Inform user script is abandoned & refuse to run
Given what I have found about the UE file formats I am not keen on
continuing to work on extract_unreal_shaders and am abandoning it. Add a
blurb to the tool to tell any potential user this, and require a lengthy
command line parameter to acknowledge this fact if they wish to use it
anyway.