generic_shader_extractor: Drop use of mmap and write a stream search routine
I used mmap as a convenient way to search the files without having to
read them into memory first, but mmap still requires that there is room
in the memory map for the whole file, which is problematic when using a
32bit version of Python as that will be limited to 4GB in the best case
scenario.

Drop mmap support and switch to searching the file stream one block at a
time, aligning to the page boundaries to minimise context switches to
the kernel and maximise speed.