ddsinfo: Use frombytes() if PIL does not implement fromstring()
Seems new versions of Pillow have dropped fromstring() and throw this
error:

NotImplementedError: fromstring() has been removed. Please call
frombytes() instead.

Switch to frombytes() if fromstring() throws a NotImplementedError. We
might be able to switch unconditionally, but I'd need to check which
version added frombytes(). Note that this is not tested yet.

Reported-By: DJ-RK