Proof of Concept for a stereo injection matrix
This is a proof of concept to add a stereo correction to any projection,
view-projection or model-view-projection matrix. The idea here is to
find a way to inject a stereo correction equivalent to the nvidia
formula into a shader without (necessarily) having to modify the shader
code, and in a way that doesn't require a pure projection matrix (though
for this first proof of concept, MGSVTPP was chosen because it does use
a pure projection matrix, so we could verify things one step at a time).

It works by finding a matrix that is equivalent to the inverse
projection matrix multiplied by a projection matrix with a built in
stereo correction with an equivalent result to the nvidia formula.

The maths to find this matrix turns out to be relatively simple, however
to create it we do need to know some values from the original projection
matrix that we might not have - the near and far clipping planes. It
determines these by passing a homogeneous coordinate with Z=0 and Z=1
through the inverse of whatever matrix we do have, normalising it then
passing it through the forwards matrix. The result will have the near or
far clipping plane in the W component.

Currently this stores matrices for both eyes in a dedicated buffer,
which requires the destination shader to be modified to select the
correct one, however with some new features in 3DMigoto it should be
possible to modify a stereo copy of the original constant buffer to
allow it to all just work.