Proof of concept injecting into an inverse-view-projection matrix
This is where things start to get interesting.

This uses the same technique as the previous two examples, but this time
rather than injecting a stereo correction into a composite forwards
projection matrix, we are injecting one into a composite inverse
view-rotation-projection matrix.

The maths is slightly different since this is the inverse case, but not
actually by that much. The injection matrix turns out to be almost
identical to the forwards variant, just with some fields negated,
finding the near and far clipping planes switches the matrices around
(since we were passed the inverse instead of the forwards), and the
matrix multiplication now goes on the other side since we need to modify
the start of the composite matrix instead of the end.

This proof of concept demonstrates how this technique could be used to
fix shadows without editing the target shader. The edits that we do are
just to select the correct matrix for each eye, and could be eliminated
with some new features in 3DMigoto.