Add a custom shader to show the contents of a constant buffer live
This adds a custom shader, font and the corresponding d3dx.ini sections
to show the contents of a constant buffer live while playing the game.

It uses a geometry shader to format the values in the buffer as floating
point and generate the text, which is passed to the pixel shader to draw
the font. The vertex shader is used to invoke the geometry shader for
each entry in the constant buffer that is to be examined.

The floating point values will not be as accurate as those from libc,
but they should be close enough for the intended purpose. Frame analysis
can of course dump out the constant buffers for a more accurate
analysis, but this is intended to quickly see which values change while
e.g. moving or rotating the camera, highlighting possible view matrices
or composites.