Use Stereo Injection Technique to fix a common problem in ABZU (UE4)
This is a variant of the stereo injection technique tailored to UE4 -
specifically to fix a very common alignment issue in ABZU.

This engine introduces a few complications - it uses a "reverse Z
projection", which means that the near and far calculations wind up with
a divide by zero which propagates a NaN result through the calculations.
However, since we have access to the projection matrix we can derive the
stereo injection matrix from first principles.

The second complication is that we are modifying a composite matrix that
does not begin with an inverse projection matrix - it begins with a
resolution divide and viewport offset. We use some maths to pull out the
extra matrix so we can inject the inverse stereo injection matrix just
before the inverse projection matrix.

For now this only modifies that one matrix, but there may well be other
matrices that would be worthwhile modifying as well, but the next step
will be to try to create a stereoised constant buffer such that the
driver will pass the correct buffer to each eye when invoking the
shaders.