ABZU: Move specular highlights to correct depth
The same technique that works in Unity and MGSV to move specular
highlights to the accurate depth works in UE4 as well - subtracting
separation*convergence from the world space camera position.

Similar to Unity, this only fixes some of the specular highlights on the
shiny surface, suggesting that some of the highlights are drawn in the
forwards pass (and if so, would require all relevant shaders to be
modified, or if I can get the constant buffer modification technique to
work modifying it there might fix all specular highlights at once).

However, the shiny surface I was looking at was a glitch - the shader
had seemingly been replaced by a completely unrelated shader. Dumping
it with 3DMigoto restored it to the correct shader, but the broken one
returned with show_original and by removing the replaced shader. I'm not
sure where this came from (it could be a 3DMigoto bug, or possibly even
a 3D Vision bug), but I have observed some strange things related to
shader management in UE4 games before that this might be related to.