ABZU: Fix deferred lighting
This seems to require a slightly different pattern to most of the
shaders in this game - here we are correcting a world space coordinate.
Interestingly, the game converts that coordinate to clip space, but
attempting to adjust it after that point is not quite accurate -
possibly because it only does a 3x4 matrix multiplication.

There is definitely something very strange going on with 3DMigoto
tracking the shaders in this engine - originally it dumped out a shader
that had clearly been decompiled, recompiled and decompiled again, which
was apparent due to the immense differences compared to the version in
ShaderCache.