ABZU: Fix judder caused by temporal AA velocity buffer
3/4 of the way through the game it suddenly enables temporal AA for
whatever reason, and in 3D Vision it causes judder and wrecks grass and
other effects.

Because of the way that temporal AA works we can't even just copy the
input to the output - that does fix the grass, but since each frame is
rendered a subpixel off from the last it leaves the judder.

It seems to be a problem with the pixel velocity causing it to sample
from the wrong location, presumably because it doesn't take the stereo
correction into account when calculating the offset. Killing the
velocity looks fine to me, so that will do for now.