ABZU: Third attempt to fix temporal AA, with better results
This gives up on trying to disable temporal AA as that was causing just
as many problems than it solved, like smearing the image when the camera
was moving too fast. This time, adjust the velocity offset based on the
depth to realign the AA samples, eliminating judder and improving the
image quality considerably.