ABZU: 4th attempt to fix temporal AA
After the 3rd attempt, the particulate matter was different in each eye,
as one was being blurred out and the other seemed to be enhanced. This
adds an adjustment around the ClipToPrevClip reprojection to try to
account for this. This has solved the immediate issue, but I doubt it is
accurate as the particulate matter seems more dense than it is in 2D.
Still, it looks fine so this will do.