ABZU: Tweak the temporal AA correction to fix perceived depth artefact
A little further on in the game the fixed temporal AA made a gate look
slightly closer than it actually was. This was because the divide by 4
fudge factor was not correct - the actual value is in the shader
already, so move the fix down to work on the scaled value.