ABZU: Fix environment reflection (RISKY - need to check if this breaks anything else)
This shader I had already fixed turned up again in a hidden area with an
environment map reflection. This adds a fix for that reflection, but
potentially breaks the previous fix - I'll need to go back and check
what else it was used for and make sure they are ok.