ABZU: Fix odd clipping at water edge in algae temple
The algae temple is one of the few places we can actually see the water
edge, and from certain angles it clips in 3D in an odd curved manner.
This requires a similar fix to the common pattern, but there are two
complications:

The shader only uses 3x4 of the SVPositionToTranslatedWorld matrix,
which does not give us the right depth value. Perform this multiply a
second time with the full matrix to get the depth (we could possibly
replace the existing multiply, but I didn't want to risk it in case that
changed the X & Y values as well and I'm not about to replay the whole
game).

The constant buffer we have been using to get the render target size
(which may be different from the resolution) is not bound. We could copy
it, but since we identified the matrices in the UE4 source code we can
do something smarter and use two different matrices to achieve our goal
without having to know the resolution, and saving a matrix inverse. It's
probably not a bad idea to switch all the other shaders using this
pattern to this new one.